requirements.txt
dash==2.14.1
pandas==2.1.3
sqlalchemy==2.0.2
python_dotenv==1.0.0
sqlalchemy==2.0.2
load_dotenv==0.1.0

# dash-core-components==2.0.0
# dash-html-components==2.0.0
# dash-table==5.0.0
# plotly==5.3.1